refactor(review): isolate live event projection
