+++ tests/test_guard_oauth_reconnect.py
@@ -329,32 +329,17 @@ def test_storage_repair_recovers_missing_oauth_binding_and_claims_unowned_reques
         ),
         "2026-06-01T00:00:00+00:00",
     )
-
     result = guard_runner_module.repair_guard_cloud_connect_storage(store)
-
-    credentials = store.get_oauth_local_credentials(allow_primary=True)
-    assert credentials is not None
+    credentials = store.get_oauth_local_credentials(allow_primary=True) or pytest.fail("missing OAuth credentials")
     assert credentials["grant_id"] == "grant-1"
-    assert result["repaired_oauth_binding"] is True
-    assert result["claimed_live_requests"] == 0
-    binding = store.get_live_request_oauth_binding()
-    assert binding is not None
-    assert store.live_request_outbox_status(now="9999-12-31T23:59:59+00:00")["quarantined_depth"] == 1
-    assert (
-        store.reassign_quarantined_live_request_outbox(
-            approved_source="default",
-            approved_workspace_id="workspace-1",
-        )
-        == 1
-    )
-    rows = store.list_ready_live_request_outbox(
-        now="9999-12-31T23:59:59+00:00",
-        limit=10,
-        oauth_subject_hash=binding["oauth_subject_hash"],
-        workspace_id=binding["workspace_id"],
-        machine_id=binding["machine_id"],
-        machine_installation_id=binding["machine_installation_id"],
-    )
+    assert result["repaired_oauth_binding"] is True and result["claimed_live_requests"] == 0
+    binding = store.get_live_request_oauth_binding() or pytest.fail("missing live-request binding")
+    late = "9999-12-31T23:59:59+00:00"
+    assert store.live_request_outbox_status(now=late)["quarantined_depth"] == 1
+    approved = {"approved_source": "default", "approved_workspace_id": "workspace-1"}
+    assert store.reassign_quarantined_live_request_outbox(**approved) == 1
+    delivery = {key: value for key, value in binding.items() if key != "oauth_source"}
+    rows = store.list_ready_live_request_outbox(now=late, limit=10, **delivery)
     assert rows[0]["oauth_subject_hash"] == binding["oauth_subject_hash"]
